+++ TowerDefence-SharedContent/Soldiers/Soldier.cs
@@ -11,8 +11,8 @@ namespace TowerDefence_SharedContent.Soldiers
         public int[] UpgradePrice { get; set; }
         public int[] BuyPrice { get; set; }
         public double Speed { get; set; }
-        public double[] Hitpoints { get; set; }
-        public double CurrentHitpoints { get; set; }
+        public int[] Hitpoints { get; set; }
+        public int CurrentHitpoints { get; set; }
         public Point Coordinates { get; set; }
         public string Sprite { get; set; }
         public SoldierType SoldierType { get; set; }
